dbus: Allow root to own the name on the bus
Distributions are supposed to run the service as special user and when
they do, it means root can't run/launch geoclue. That is annoying for
bug reporters who need to launch geoclue manually for providing console
log from it.

Let's ensure that root is always able to launch the service.