Add GClueNMEASource
GClueNMEASource receives NMEA GGA sentences from a NMEA source on the
network. It uses Avahi for the auto-discovery of nmea-tcp services on
local network.

This will enable us to receive location info from other devices on the
local network with location capabilities, such as smartphones.

This patch adds dependency on avahi-glib and avahi-client.

https://bugs.freedesktop.org/show_bug.cgi?id=90974