Workaround an intltool issue
`make distcheck` breaks with the error:

-----
The following files contain translations and are currently not in use.
Please
consider adding these to the POTFILES.in file, located in the po/
directory.

sub/demo/geoclue-demo-agent.desktop.in
sub/demo/geoclue-where-am-i.desktop.in
----

The only known workaround is to add these paths to POTFILES.skip and
this patch does that.