nmea-source: Ensure 1st service remains 1st in list
Currently if a second service is discovered that has the same accuracy
as the first one in the list, g_list_insert_sorted() can (and in
practice does) insert it in the beginning since our comparison function
declares them both equal. Because of this, we end up switching to the
new service even if its accuracy level is the same as the one we are
already using.

Let's solve this by also taking time of discovery into account in our
comparison function.