location: Only ignore timestamps > 60s in future
If timestamp is from a remote machine, we could easily end up ignoring
the timestamp as future timestamp (and ultimately ignoring the
associated location update) if clocks on two machines are out of sync.

Let's solve this by only ignoring the timestamp if it's at least 60
seconds in the future.