client-info: Don't load props of /org/freedesktop/DBus
By default, glib's GDBusProxy calls org.freedesktop.DBus.Properties.GetAll
to load all properties on a loaded interface - however in this case we
are creating a proxy to the system standard bus interface for the sole
purpose of later calling GetConnectionUnixUser.

GeoClue does not have permission to use the org.freedesktop.DBus.Properties
interface on a proxy to /org/freedesktop/DBus (and there are no properties
to be loaded anyway).

Note that GeoClue works fine without this patch, however this patch
will avoid error messages being logged to the system logs.

https://bugs.freedesktop.org/show_bug.cgi?id=92979