web-source: Check for full internet availability
Instead of just checking if some network is available, let's check if we
are fully connected to internet or not.

If glib >= 2.44 is available, we make use the new API added in that
version to check for connectivity but if it's not, we check if Mozilla
Location Services hostname is reachable.

https://bugs.freedesktop.org/show_bug.cgi?id=95493